examples/legal-agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.9.2
python-dotenv>=1.0.1
langchain>=0.3.0
qdrant-client>=1.11.0
pytest>=8.3.3
httpx>=0.27.2
